pyproject.toml
[tool.poetry]
name = "gelendarus-bot"
version = "0.1.0"
description = ""
authors = ["sebastianhamza <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.0.1"
Flask = "^2.0.1"
qrcode = "^7.2"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
